Keep the redirect stubs within the title limit, and gate the map in CI
Astro titles a redirect stub with the destination spelled out in full, which
runs past the 70-character limit html-validate holds every other page to as
soon as the address is more than four segments deep. With eight shallow stubs
nothing showed; with one per published address, 66 titles were over. The stub
is noindex and on screen for one frame, and its destination is still in the
meta refresh, the canonical link and the visible link in the body, so the title
becomes plain "Redirecting". The postbuild script that was already normalising
the doctype of those same pages now does both, and is named for the pages it
fixes rather than for one of the two fixes.

The docs workflow runs the redirect check after the build: it derives the map
from the rename history again and fails if the committed one has fallen behind,
then confirms against dist that no redirect points at a missing page, at
another redirect, or at an address a real page already occupies.
